Created calculateSentimentScore() function
Currently, it takes a sentence and the lexicon dictionary (and it's size) and it states if each word in the sentence is in the dictionary.

Also updated the createLexiconDictionary() function to take a pointer to an int as a parameter (n), and saves the number of lines in the file (aka the size of the lexicon dictionary) to the pointer.