Added ALL CAPS to calculation
*IMPORTANT*
Came across an issue in main() that would cause a slightly longer runtime and not print anything to the terminal if a for loop that was used in testing is removed or commented out.
As I was developing the ALL CAPS implementation, a similar issue arose with a print statement used for testing, and now removing that print statement causes the same issue.

Will prioritize fixing this issue for now.