Updated createSentimentScore() and fixed hidden bugs
There was a hidden segmentation fault if the sentence reached a certain length since the memory allocated for upperWord wasn't changing as the word changed. It would only stay as the length of the first word. This was fixed by dynamically allocating (and freeing) upperWord within the main loop.

Also learned that sentences ending with a period would mess up the sentiment calculation since the last word wouldn't be found in the dictionary. Temporarily solved this by removing periods along with exclamation marks. However, this is not a long term solution, as some of the "words" in the dictionary have periods (and actually, commas as well), so those "word's" values wouldn't be considered in the sentiment score.
This is acceptable for now, as these "words" are very rarely used, but a proper fix to this should be found in the long-term.