Fixed issue of code not being able to compile in Linux, and configured Makefile
According to Linux, there were multiple instances of the amplifier lists being defined. This was fixed by declaring the amplifiers in utility.h using the extern keyword, which essentially tells the compiler that the variable exists somewhere else.

There was also an error saying that there was an undefined reference to the sqrt() function. This was fixed by adding the -lm flag when compiling (see in Makefile).

Also added .out files and the main executable file to .gitignore